+++ rust/operator-binary/src/connect/crd.rs
@@ -81,7 +81,6 @@ type SparkConnectExecutorRoleType = CommonConfiguration<
     )
 )]
 pub mod versioned {
-
     /// An Apache Spark Connect server component. This resource is managed by the Stackable operator
     /// for Apache Spark. Find more information on how to use it in the
     /// [operator documentation](DOCS_BASE_URL_PLACEHOLDER/spark-k8s/usage-guide/connect-server).