+++ rust/operator-binary/src/connect/server.rs
@@ -156,12 +156,9 @@ pub(crate) fn server_config_map(
 ) -> Result<ConfigMap, Error> {
     let cm_name = object_name(&scs.name_any(), SparkConnectRole::Server);
 
-    let security_properties_overrides = scs
-        .spec
-        .server
-        .config
-        .as_ref()
-        .map(|s| &s.config_overrides)
+    let config_overrides = scs.spec.server.config.as_ref().map(|s| &s.config_overrides);
+
+    let security_properties_overrides = config_overrides
         .and_then(|config_overrides| config_overrides.security_properties.as_ref())
         .map(KeyValueConfigOverrides::as_product_config_overrides)
         .unwrap_or_default();
@@ -172,12 +169,8 @@ pub(crate) fn server_config_map(
         },
     )?;
 
-    let metrics_properties_overrides = scs
-        .spec
-        .server
-        .config
-        .as_ref()
-        .and_then(|s| s.config_overrides.metrics_properties.as_ref())
+    let metrics_properties_overrides = config_overrides
+        .and_then(|config_overrides| config_overrides.metrics_properties.as_ref())
         .map(KeyValueConfigOverrides::as_product_config_overrides)
         .unwrap_or_default();